+++ assets/imgui.ini
@@ -1,0 +1,17 @@
+[Window][Debug##Default]
+Pos=60,60
+Size=400,400
+Collapsed=0
+
+[Window][Dear ImGui Demo]
+Pos=650,20
+Size=550,680
+Collapsed=1
+
+[Window][controlls]
+Pos=60,60
+Size=128,157
+Collapsed=0
+
+[Docking][Data]
+
