+++ utils/quick_math.cpp
@@ -1,4 +1,4 @@
-#include "math.h"
+#include "quick_math.h"
 #include <glm/glm.hpp>
 
 namespace math {
@@ -17,14 +17,11 @@ ChangeOfBasis(const glm::vec3& point, const glm::vec3& xAxis, const glm::vec3& y
     glm::vec3 normYAxis = glm::normalize(yAxis);
     glm::vec3 normZAxis = glm::normalize(zAxis);
 
-    glm::mat4 rotation = glm::mat4(glm::vec4(normXAxis, 0.0f),
-                                   glm::vec4(normYAxis, 0.0f), // NOTE: ZAxis is the new YAxis
-                                   glm::vec4(normZAxis, 0.0f),
-                                   glm::vec4(0.0f, 0.0f, 0.0f, 1.0f));
-    glm::mat4 transform = glm::inverse(rotation);
+    glm::mat3 rotation = glm::mat3(normXAxis, normYAxis, normZAxis);
+    glm::mat3 transform = glm::inverse(rotation);
 
-    glm::vec4 transformedPoint = transform * glm::vec4(point, 1.0f);
+    glm::vec3 transformedPoint = transform * point;
 
-    return glm::vec3(transformedPoint);
+    return transformedPoint;
 }
 } // namespace math
