+++ utils/quick_math.h
